greek-edu-rag/requirements.txt
# ── Core FastAPI stack ────────────────────────────────────────
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
pydantic>=2.6,<3.0
PyJWT>=2.8                         # JWT για Supabase auth (HS256)
cryptography>=42.0                  # απαιτείται από PyJWT

# ── Supabase + Auth ───────────────────────────────────────────
supabase>=2.4,<3.0

# ── LLM — Anthropic Claude ────────────────────────────────────
anthropic>=0.25

# ── RAG (Pinecone) ────────────────────────────────────────────
# Προσθέτουμε αργότερα όταν στηθεί το Pinecone account
# pinecone>=3.0,<4.0

# ── Emails (Resend) ───────────────────────────────────────────
resend>=0.7

# ── Payments (Stripe) ─────────────────────────────────────────
stripe>=9.0

# ── Utils ─────────────────────────────────────────────────────
python-dateutil>=2.8
httpx>=0.27
python-dotenv>=1.0
json-repair>=0.30                   # επιδιόρθωση malformed LLM JSON
